llm_consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
google-generativeai==0.7.1
prometheus-client==0.17.1